+++ apps/intellij/src/main/kotlin/dev/nx/console/nxls/server/NxlsLanguageServer.kt
@@ -1,13 +1,13 @@
-package dev.nx.console.lsp.server
+package dev.nx.console.nxls.server
 
 import org.eclipse.lsp4j.jsonrpc.services.JsonDelegate
 import org.eclipse.lsp4j.services.LanguageServer
 
 interface NxlsLanguageServer : LanguageServer {
-  @JsonDelegate
-  fun getNxService(): NxService {
-    throw UnsupportedOperationException()
-  }
+    @JsonDelegate
+    fun getNxService(): NxService {
+        throw UnsupportedOperationException()
+    }
 
 }
